Add review comment changes
