+++ java/java.hints/src/org/netbeans/modules/java/hints/jdk/mapreduce/TreeUtilities.java
@@ -21,39 +21,7 @@
  */
 package org.netbeans.modules.java.hints.jdk.mapreduce;
 
-import com.sun.source.tree.AssignmentTree;
-import com.sun.source.tree.BlockTree;
-import com.sun.source.tree.BreakTree;
-import com.sun.source.tree.CaseTree;
-import com.sun.source.tree.ClassTree;
-import com.sun.source.tree.ContinueTree;
-import com.sun.source.tree.ExpressionStatementTree;
-import com.sun.source.tree.ExpressionTree;
-import com.sun.source.tree.IdentifierTree;
-import com.sun.source.tree.IfTree;
-import com.sun.source.tree.LambdaExpressionTree;
-import com.sun.source.tree.ReturnTree;
-import com.sun.source.tree.StatementTree;
-import com.sun.source.tree.SwitchTree;
-import com.sun.source.tree.ThrowTree;
 import com.sun.source.tree.Tree;
-import com.sun.source.util.TreePath;
-import com.sun.source.util.TreePathScanner;
-import com.sun.tools.javac.tree.JCTree;
-import java.util.ArrayList;
-import java.util.EnumSet;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Set;
-import javax.lang.model.element.Element;
-import javax.lang.model.element.Modifier;
-import javax.lang.model.element.VariableElement;
-import org.netbeans.api.java.source.CompilationInfo;
-import org.netbeans.api.java.source.TreeMaker;
-import org.netbeans.api.java.source.WorkingCopy;
-import org.netbeans.modules.java.source.TreeShims;
-import org.netbeans.spi.java.hints.JavaFix;
 
 /**
  *
@@ -80,196 +48,4 @@ public class TreeUtilities {
                 || opKind == Tree.Kind.POSTFIX_DECREMENT
                 || opKind == Tree.Kind.PREFIX_DECREMENT;
     }
-
-    public static boolean completesNormally(CompilationInfo info, TreePath tp) {
-        class Scanner extends TreePathScanner<Void, Void> {
-
-            private boolean completesNormally = true;
-            private Set<Tree> seenTrees = new HashSet<>();
-
-            @Override
-            public Void visitReturn(ReturnTree node, Void p) {
-                completesNormally = false;
-                return null;
-            }
-
-            @Override
-            public Void visitBreak(BreakTree node, Void p) {
-                completesNormally &= seenTrees.contains(info.getTreeUtilities().getBreakContinueTarget(getCurrentPath()));
-                return null;
-            }
-
-            @Override
-            public Void visitContinue(ContinueTree node, Void p) {
-                completesNormally &= seenTrees.contains(info.getTreeUtilities().getBreakContinueTarget(getCurrentPath()));
-                return null;
-            }
-
-            @Override
-            public Void visitThrow(ThrowTree node, Void p) {
-                completesNormally = false;
-                return null;
-            }
-
-            @Override
-            public Void visitIf(IfTree node, Void p) {
-                boolean origCompletesNormally = completesNormally;
-                scan(node.getThenStatement(), p);
-                boolean afterThen = completesNormally;
-                completesNormally = origCompletesNormally;
-                scan(node.getElseStatement(), p);
-                completesNormally |= afterThen;
-                return null;
-            }
-
-            @Override
-            public Void visitSwitch(SwitchTree node, Void p) {
-                //exhaustiveness: (TODO)
-                boolean hasDefault = node.getCases().stream().anyMatch(c -> c.getExpression() == null);
-                if (node.getCases().size() > 0) {
-                    scan(node.getCases().get(node.getCases().size() - 1), p);
-                }
-                completesNormally |= !hasDefault;
-                return null;
-            }
-
-            //TODO: loops
-            @Override
-            public Void scan(Tree tree, Void p) {
-                seenTrees.add(tree);
-                return super.scan(tree, p);
-            }
-
-            @Override
-            public Void visitLambdaExpression(LambdaExpressionTree node, Void p) {
-                return null;
-            }
-
-            @Override
-            public Void visitClass(ClassTree node, Void p) {
-                return null;
-            }
-        }
-
-        Scanner scanner = new Scanner();
-
-        scanner.scan(tp, null);
-        return scanner.completesNormally;
-    }
-
-    public static void performRewriteRuleSwitch(JavaFix.TransformationContext ctx, TreePath tp, Tree st) {
-        WorkingCopy wc = ctx.getWorkingCopy();
-        TreeMaker make = wc.getTreeMaker();
-        List<CaseTree> newCases = new ArrayList<>();
-        List<? extends CaseTree> cases;
-        Set<VariableElement> variablesDeclaredInOtherCases = new HashSet<>();
-        List<ExpressionTree> patterns = new ArrayList<>();
-        boolean switchExpressionFlag = st.getKind().toString().equals("SWITCH_EXPRESSION");
-        if(switchExpressionFlag){
-            cases = TreeShims.getCases(st);
-        }else{
-            cases = ((SwitchTree)st).getCases();
-        }
-        for (Iterator<? extends CaseTree> it = cases.iterator(); it.hasNext();) {
-            CaseTree ct = it.next();
-            TreePath casePath = new TreePath(tp, ct);
-            patterns.addAll(TreeShims.getExpressions(ct));
-            List<StatementTree> statements;
-            if (ct.getStatements() == null) {
-                statements = new ArrayList<>(((JCTree.JCCase) ct).stats);
-            } else {
-                statements = new ArrayList<>(ct.getStatements());
-            }
-            if (statements.isEmpty()) {
-                if (it.hasNext()) {
-                    continue;
-                }
-                //last case, no break
-            } else if (!switchExpressionFlag && statements.get(statements.size() - 1).getKind() == Tree.Kind.BREAK
-                    && ctx.getWorkingCopy().getTreeUtilities().getBreakContinueTarget(new TreePath(new TreePath(tp, ct), statements.get(statements.size() - 1))) == st) {
-                statements.remove(statements.size() - 1);
-            } else {
-                new TreePathScanner<Void, Void>() {
-                    @Override
-                    public Void visitBlock(BlockTree node, Void p) {
-                        if (!node.getStatements().isEmpty()
-                                && node.getStatements().get(node.getStatements().size() - 1).getKind() == Tree.Kind.BREAK
-                                && ctx.getWorkingCopy().getTreeUtilities().getBreakContinueTarget(new TreePath(getCurrentPath(), node.getStatements().get(node.getStatements().size() - 1))) == st) {
-                            wc.rewrite(node, make.removeBlockStatement(node, node.getStatements().get(node.getStatements().size() - 1)));
-                            //TODO: optimize ifs?
-                        }
-                        return super.visitBlock(node, p);
-                    }
-                }.scan(new TreePath(new TreePath(tp, ct), statements.get(statements.size() - 1)), null);
-            }
-            Set<Element> seenVariables = new HashSet<>();
-            int idx = 0;
-            for (StatementTree statement : new ArrayList<>(statements)) {
-                TreePath statementPath = new TreePath(casePath, statement);
-                if (statement.getKind() == Tree.Kind.EXPRESSION_STATEMENT) {
-                    ExpressionTree expr = ((ExpressionStatementTree) statement).getExpression();
-                    if (expr.getKind() == Tree.Kind.ASSIGNMENT) {
-                        AssignmentTree at = (AssignmentTree) expr;
-                        Element var = wc.getTrees().getElement(new TreePath(new TreePath(statementPath, at), at.getVariable()));
-                        if (variablesDeclaredInOtherCases.contains(var)) {
-                            seenVariables.add(var);
-                            //XXX: take type from the original variable
-                            wc.rewrite(statement,
-                                    make.Variable(make.Modifiers(EnumSet.noneOf(Modifier.class)), var.getSimpleName(), make.Type(var.asType()), at.getExpression()));
-                        }
-                    }
-                }
-                Set<Element> thisStatementSeenVariables = new HashSet<>();
-                new TreePathScanner<Void, Void>() {
-                    @Override
-                    public Void visitIdentifier(IdentifierTree node, Void p) {
-                        Element el = wc.getTrees().getElement(getCurrentPath());
-                        if (variablesDeclaredInOtherCases.contains(el) && seenVariables.add(el)) {
-                            thisStatementSeenVariables.add(el);
-                        }
-                        return super.visitIdentifier(node, p);
-                    }
-                }.scan(statementPath, null);
-
-                if (!thisStatementSeenVariables.isEmpty()) {
-                    for (Element el : thisStatementSeenVariables) {
-                        VariableElement var = (VariableElement) el;
-                        statements.add(idx++, make.Variable(make.Modifiers(EnumSet.noneOf(Modifier.class)), var.getSimpleName(), make.Type(var.asType()), null));
-                    }
-                }
-                idx++;
-            }
-            Tree body = make.Block(statements, false);
-            if (statements.size() == 1) {
-                if (statements.get(0).getKind() == Tree.Kind.EXPRESSION_STATEMENT
-                        || statements.get(0).getKind() == Tree.Kind.THROW
-                        || statements.get(0).getKind() == Tree.Kind.BLOCK) {
-                    body = statements.get(0);
-                }
-            }
-            newCases.add(make.Case(patterns, body));
-            patterns = new ArrayList<>();
-            for (StatementTree statement : getSwitchStatement(ct)) {
-                if (statement.getKind() == Tree.Kind.VARIABLE) {
-                    variablesDeclaredInOtherCases.add((VariableElement) wc.getTrees().getElement(new TreePath(casePath, statement)));
-                }
-            }
-        }
-        if(switchExpressionFlag){
-            wc.rewrite(st, make.SwitchExpression(TreeShims.getExpressions(st).get(0), newCases));
-        }else{
-            wc.rewrite((SwitchTree)st, make.Switch(((SwitchTree)st).getExpression(), newCases));
-        }
-    }
-
-    private static List<? extends StatementTree> getSwitchStatement(CaseTree ct) {
-        if (ct.getStatements() != null) {
-            return ct.getStatements();
-        } else if (ct instanceof JCTree.JCCase) {
-            return ((JCTree.JCCase) ct).stats;
-        } else {
-            return null;
-        }
-    }
-
 }
