+++ java/java.source.base/apichanges.xml
@@ -25,6 +25,18 @@
     <apidef name="javasource_base">Java Source API</apidef>
 </apidefs>
 <changes>
+    <change id="TreeMaker.SwitchExpression">
+        <api name="javasource_base" />
+        <summary>Creates a new SwitchExpressionTree</summary>
+        <version major="1" minor="2.41.0"/>
+        <date day="27" month="2" year="2019"/>
+        <author login="vikasprabhakar"/>
+        <compatibility addition="yes" binary="compatible" source="compatible"/>
+        <description>
+                Creates a new SwitchExpressionTree.
+        </description>
+        <class name="TreeMaker" package="org.netbeans.api.java.source"/>
+    </change>
     <change id="Matcher.setKeepSyntheticTrees">
         <api name="javasource_base"/>
         <summary>Adding Matcher.setKeepSyntheticTrees method</summary>
